Updated script.js token creation
latest teams update removed "latestOid" from the ts.x.cahce field. Workaround was to pull the new id from a ts.openDbs field, remove irrelevant text and use it in place of where latestOid was in token creation.

I'm not sure if the openDbs field will list different fields based on the what a company has or has not subscribed to for the enterprise, however this is what is working for my setup.